// src/types/graphql.d.ts
export type Maybe<T> = T | null;
export type FieldWrapper<T> = T;
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type IChannel = {
  id: FieldWrapper<Scalars['ID']>;
  name: FieldWrapper<Scalars['String']>;
  isFavorite: FieldWrapper<Scalars['Boolean']>;
  messages?: Maybe<Array<FieldWrapper<IMessage>>>;
};

export type ICreateChannelInput = {
  name: Scalars['String'];
};

export type IFavoriteChannelInput = {
  id: Scalars['ID'];
  isFavorite: Scalars['Boolean'];
};

export type IMessage = {
  id: FieldWrapper<Scalars['ID']>;
  type: FieldWrapper<Scalars['String']>;
  createdAt: FieldWrapper<Scalars['String']>;
};